Fix Discovery v5 implementation based on code review feedback
- Add isInitiator flag to ActiveSession for proper role tracking
- Fix encryption/decryption key selection based on session role
- Fix ephemeral public key generation using proper EC point multiplication
- Add signature generation over id-nonce in handshake
- Fix WHOAREYOU nonce to echo the triggering packet's nonce
- Fix handshake correlation using nonce-based tracking
- Improve node ID extraction from packets
- Add fiber tracking and cleanup in startHandling
- Fix NODES codec to use uint16 for total field
- Fix FINDNODE response handling
- Add localEnrSeq parameter to DiscoveryNetwork.apply

+++ scalanet/discovery/src/com/chipprbots/scalanet/discovery/ethereum/v5/codecs/RLPCodecs.scala
@@ -78,7 +78,7 @@ trait PayloadCodecs {
   
   /** Codec for NODES message (0x04) */
   given nodesCodec: Codec[Nodes] = {
-    (requestIdCodec :: uint8 :: listOfN(uint8, enrCodec)).xmap(
+    (requestIdCodec :: uint16 :: listOfN(uint16, enrCodec)).xmap(
       { case (reqId, total, enrs) => Nodes(reqId, total, enrs) },
       (nodes: Nodes) => (nodes.requestId, nodes.total, nodes.enrs)
     )
